@@ -14,15 +14,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.10",
-    "@angular/common": "^6.1.10",
-    "@angular/compiler": "^6.1.10",
-    "@angular/core": "^6.1.10",
-    "@angular/forms": "^6.1.10",
+    "@angular/animations": "^11.0.0",
+    "@angular/common": "^11.0.0",
+    "@angular/compiler": "^11.0.0",
+    "@angular/core": "^11.0.0",
+    "@angular/forms": "^11.0.0",
     "@angular/http": "^6.1.10",
-    "@angular/platform-browser": "^6.1.10",
-    "@angular/platform-browser-dynamic": "^6.1.10",
-    "@angular/router": "^6.1.10",
+    "@angular/platform-browser": "^11.0.0",
+    "@angular/platform-browser-dynamic": "^11.0.0",
+    "@angular/router": "^11.0.0",
     "angular-cli-ghpages": "^0.6.0",
     "core-js": "^2.6.5",
     "handlebars": "^4.7.6",
@@ -30,16 +30,16 @@
     "jquery": "^3.5.0",
     "ngx-toastr": "^10.0.2",
     "querystringify": "^2.0.0",
-    "rxjs": "^6.5.1",
+    "rxjs": "^6.6.3",
     "rxjs-compat": "^6.5.1",
     "sweetalert": "^2.1.2",
-    "zone.js": "^0.8.29"
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.1100.1",
     "@angular/cli": "^11.0.1",
-    "@angular/compiler-cli": "^7.2.14",
-    "@angular/language-service": "^6.1.10",
+    "@angular/compiler-cli": "^11.0.0",
+    "@angular/language-service": "^11.0.0",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "^6.0.118",
@@ -55,6 +55,6 @@
     "protractor": "^5.4.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "^3.2.1"
+    "typescript": "^4.0.5"
   }
 }
