@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { TrapezoideComponent } from './trapezoide.component';
 
@@ -6,7 +6,7 @@ describe('TrapezoideComponent', () => {
   let component: TrapezoideComponent;
   let fixture: ComponentFixture<TrapezoideComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ TrapezoideComponent ]
     })
