@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { AptitudesComponent } from './aptitudes.component';
 
@@ -6,7 +6,7 @@ describe('AptitudesComponent', () => {
   let component: AptitudesComponent;
   let fixture: ComponentFixture<AptitudesComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ AptitudesComponent ]
     })
