@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ProyectosComponent } from './proyectos.component';
 
@@ -6,7 +6,7 @@ describe('ProyectosComponent', () => {
   let component: ProyectosComponent;
   let fixture: ComponentFixture<ProyectosComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ProyectosComponent ]
     })
