@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ModalProyectosComponent } from './modal-proyectos.component';
 
@@ -6,7 +6,7 @@ describe('ModalProyectosComponent', () => {
   let component: ModalProyectosComponent;
   let fixture: ComponentFixture<ModalProyectosComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ModalProyectosComponent ]
     })
